+++ firmware/common/portapack_persistent_memory.hpp
@@ -46,8 +46,8 @@ void set_afsk_space(const int32_t new_value);
 int32_t afsk_bitrate();
 void set_afsk_bitrate(const int32_t new_value);
 
-uint8_t afsk_config();
-void set_afsk_config(const uint8_t new_value);
+uint32_t afsk_config();
+void set_afsk_config(const uint32_t new_value);
 
 bool playing_dead();
 void set_playing_dead(const bool new_value);
